# contracts/openapi/parslex-v0.yaml
openapi: 3.1.0
info:
  title: ParsLex Platform API
  version: 1.1.0
  description: |
    Enterprise Domain-Adaptive Legal AI Platform — REST API v0.

    Phase 1 covers identity, documents, collections, assistant sessions, and audit.
    All endpoints require authentication unless noted.
  contact:
    name: ParsLex Architecture Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://parslex.local/api/v1
    description: On-premise deployment
...
paths:
  /auth/login:
    post:
      tags: [Identity]
      summary: Authenticate user
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Identity]
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /organizations/{orgId}/assistant/sessions:
    get:
      tags: [Assistant]
      summary: List AI assistant sessions
      operationId: listAssistantSessions
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: workspace_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AISession'
    post:
      tags: [Assistant]
      summary: Create AI assistant session
      operationId: createAssistantSession
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAISessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AISession'
  /organizations/{orgId}/assistant/sessions/{sessionId}:
    get:
      tags: [Assistant]
      summary: Get assistant session with messages
      operationId: getAssistantSession
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session with messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AISessionDetail'
  /organizations/{orgId}/assistant/sessions/{sessionId}/messages:
    post:
      tags: [Assistant]
      summary: Send message to assistant
      operationId: sendAssistantMessage
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIMessage'
  /organizations/{orgId}/audit/events:
    get:
      tags: [Audit]
      summary: List audit events
      operationId: listAuditEvents
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: resource_type
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Audit event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventListResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, email, full_name, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [admin, legal_user, business_user, viewer]
        created_at:
          type: string
          format: date-time
...
    AISession:
      type: object
      required: [id, organization_id, title, created_at]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        title:
          type: string
        legal_domain_id:
          type: string
        scope:
          $ref: '#/components/schemas/AISessionScope'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AISessionScope:
      type: object
      properties:
        document_ids:
          type: array
          items:
            type: string
            format: uuid
        collection_ids:
          type: array
          items:
            type: string
            format: uuid
    AISessionDetail:
      allOf:
        - $ref: '#/components/schemas/AISession'
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: '#/components/schemas/AIMessage'
    CreateAISessionRequest:
      type: object
      properties:
        title:
          type: string
        workspace_id:
          type: string
          format: uuid
        legal_domain_id:
          type: string
          default: iran-oil-gas
        scope:
          $ref: '#/components/schemas/AISessionScope'
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
        stream:
          type: boolean
          default: false
    AIMessage:
      type: object
      required: [id, session_id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        model_version:
          type: string
        created_at:
          type: string
          format: date-time
    Citation:
      type: object
      required: [document_id, chunk_id, excerpt]
      properties:
        document_id:
          type: string
          format: uuid
        document_title:
          type: string
        chunk_id:
          type: string
          format: uuid
        excerpt:
          type: string
        page_number:
          type: integer
    AuditEvent:
      type: object
      required: [id, organization_id, action, resource_type, created_at]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_email:
          type: string
        action:
          type: string
          enum:
            - login
            - logout
            - document.upload
            - document.download
            - document.delete
            - collection.create
            - collection.update
            - assistant.message
        resource_type:
          type: string
        resource_id:
          type: string
          format: uuid
        metadata:
          type: object
          additionalProperties: true
        ip_address:
          type: string
        created_at:
          type: string
          format: date-time
    AuditEventListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer